눈치게임 완성
